+++ modules/tinymce/src/core/main/ts/dom/DomSerializerFilters.ts
@@ -195,7 +195,7 @@ const register = (htmlParser: DomParser, settings: DomSerializerSettings, dom: D
   // make it possible to place the caret inside empty blocks. This logic tries to remove
   // these elements and keep br elements that where intended to be there intact
   if (settings.remove_trailing_brs) {
-    RemoveTrailingBr.addNodeFilter(htmlParser, htmlParser.schema);
+    RemoveTrailingBr.addNodeFilter(settings, htmlParser, htmlParser.schema);
   }
 };