+++ modules/tinymce/src/core/main/ts/dom/RemoveTrailingBr.ts
@@ -6,8 +6,9 @@ import Schema from '../api/html/Schema';
 import Tools from '../api/util/Tools';
 import * as TransparentElements from '../content/TransparentElements';
 import { isEmpty, paddEmptyNode } from '../html/ParserUtils';
+import { DomSerializerSettings } from './DomSerializerImpl';
 
-export const addNodeFilter = (htmlParser: DomParser, schema: Schema): void => {
+export const addNodeFilter = (settings: DomSerializerSettings, htmlParser: DomParser, schema: Schema): void => {
   htmlParser.addNodeFilter('br', (nodes, _, args) => {
     const blockElements = Tools.extend({}, schema.getBlockElements());
     const nonEmptyElements = schema.getNonEmptyElements();
@@ -16,7 +17,7 @@ export const addNodeFilter = (htmlParser: DomParser, schema: Schema): void => {
     // Remove brs from body element as well
     blockElements.body = 1;
 
-    const isBlock = (node: AstNode) => node.name in blockElements && TransparentElements.isTransparentAstInline(schema, node);
+    const isBlock = (node: AstNode) => node.name in blockElements || TransparentElements.isTransparentAstInline(schema, node);
 
     // Must loop forwards since it will otherwise remove all brs in <p>a<br><br><br></p>
     for (let i = 0, l = nodes.length; i < l; i++) {
@@ -54,7 +55,7 @@ export const addNodeFilter = (htmlParser: DomParser, schema: Schema): void => {
               if (elementRule.removeEmpty) {
                 parent.remove();
               } else if (elementRule.paddEmpty) {
-                paddEmptyNode(args, isBlock, parent);
+                paddEmptyNode(settings, args, isBlock, parent);
               }
             }
           }